+++ WebContent/test/autotest-class.js
@@ -1,22 +1,5 @@
 window.miniTests.push.apply(window.miniTests, [
   	{
-		name:'$().hasClass()',
-		exec: function() {
-			var s1, s3;
-			$('#container2').add(s1 = EE('div', {'className': 'a b c d e'})())
-					.add(EE('div', {'className': ''}, []))
-					.add(s3 = EE('div', {'className': 'a d f'}, [])());
-			var m = $('#container2 div');
-			check(m.hasClass('a'), s1[0], true);
-			check(m.hasClass('b'), s1[0], true);
-			check(m.hasClass('c'), s1[0], true);
-			check(m.hasClass('d'), s1[0], true);
-			check(m.hasClass('e'), s1[0], true);
-			check(m.hasClass('f'), s3[0], true);
-			check(!m.hasClass('x'));
-		}
-	},
-	{
 		name:'$().set("$", "-")',
 		exec: function() {
 			var s1 = EE('div', {'className': 'a b c d e'})()[0];